Refactor pagination of item cards
Adds new component for presenting items. Resolves #41.
